requirements.txt
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
defusedxml==0.6.0
Django==1.11.23
django-allauth==0.39.1
django-appconf==1.0.3
django-braces==1.13.0
django-imagekit==4.0.2
django-ipware==2.1.0
django-model-utils==3.1.2
django-notifications-hq==1.5.0
django-select2==6.1.2
idna==2.8
jsonfield==2.0.2
kiwisolver==1.0.1
kombu==4.5.0
matplotlib==3.0.3
numpy==1.16.2
oauthlib==3.0.1
pandas==0.24.2
pilkit==2.0
Pillow==5.2.0
psycopg2-binary==2.7.5
pyparsing==2.4.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
redis==3.2.1
requests==2.21.0
requests-oauthlib==1.2.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
urllib3==1.24.2
vine==1.3.0
